# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"MIRA — Medical Imaging Recognition Assistant","description":"Headless, API-first medical imaging analysis. **Research prototype — not a medical device.**","version":"0.1.0"},"paths":{"/api/v1/studies":{"post":{"tags":["studies"],"summary":"Upload Study","operationId":"upload_study_api_v1_studies_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_study_api_v1_studies_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudyPublic"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/studies/{study_id}":{"get":{"tags":["studies"],"summary":"Get Study","operationId":"get_study_api_v1_studies__study_id__get","parameters":[{"name":"study_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Study Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudyPublic"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/studies/{study_id}/image":{"get":{"tags":["studies"],"summary":"Get Study Image","operationId":"get_study_image_api_v1_studies__study_id__image_get","parameters":[{"name":"study_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Study Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/studies/{study_id}/inference":{"post":{"tags":["inference"],"summary":"Submit Inference","operationId":"submit_inference_api_v1_studies__study_id__inference_post","parameters":[{"name":"study_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Study Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InferenceRequest"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InferenceSubmitResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/inference/{inference_id}/result":{"get":{"tags":["inference"],"summary":"Get Inference Result","operationId":"get_inference_result_api_v1_inference__inference_id__result_get","parameters":[{"name":"inference_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Inference Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InferenceResultPublic"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/inference/{inference_id}/explanation":{"get":{"tags":["inference"],"summary":"Get Explanation","operationId":"get_explanation_api_v1_inference__inference_id__explanation_get","parameters":[{"name":"inference_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Inference Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models":{"get":{"tags":["models"],"summary":"List Models","operationId":"list_models_api_v1_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ModelListItem"},"type":"array","title":"Response List Models Api V1 Models Get"}}}}}}},"/api/v1/tasks/{task_id}":{"get":{"tags":["tasks"],"summary":"Get Task Status","operationId":"get_task_status_api_v1_tasks__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"Body_upload_study_api_v1_studies_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_study_api_v1_studies_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InferenceRequest":{"properties":{"tool_id":{"type":"string","title":"Tool Id","default":"pneumonia_classifier_v1"}},"type":"object","title":"InferenceRequest"},"InferenceResultPublic":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"study_id":{"type":"string","format":"uuid","title":"Study Id"},"tool_id":{"type":"string","title":"Tool Id"},"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Result"},"gradcam_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gradcam Key"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["id","study_id","tool_id","task_id","status","created_at"],"title":"InferenceResultPublic"},"InferenceSubmitResponse":{"properties":{"inference_id":{"type":"string","format":"uuid","title":"Inference Id"},"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"tool_id":{"type":"string","title":"Tool Id"}},"type":"object","required":["inference_id","task_id","status","tool_id"],"title":"InferenceSubmitResponse"},"ModelListItem":{"properties":{"tool_id":{"type":"string","title":"Tool Id"},"name":{"type":"string","title":"Name"},"version":{"type":"string","title":"Version"},"description":{"type":"string","title":"Description"},"modality":{"type":"string","title":"Modality"},"task_type":{"type":"string","title":"Task Type"},"input_shape":{"items":{"type":"integer"},"type":"array","title":"Input Shape"},"class_names":{"items":{"type":"string"},"type":"array","title":"Class Names"},"loaded":{"type":"boolean","title":"Loaded"}},"type":"object","required":["tool_id","name","version","description","modality","task_type","input_shape","class_names","loaded"],"title":"ModelListItem"},"StudyPublic":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"file_format":{"type":"string","title":"File Format"},"file_size":{"type":"integer","title":"File Size"},"anonymized":{"type":"boolean","title":"Anonymized"},"file_key":{"type":"string","title":"File Key"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"dicom_metadata":{"additionalProperties":true,"type":"object","title":"Dicom Metadata"}},"type":"object","required":["id","file_format","file_size","anonymized","file_key","created_at","dicom_metadata"],"title":"StudyPublic"},"TaskStatusResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"inference_result_id":{"type":"string","format":"uuid","title":"Inference Result Id"}},"type":"object","required":["task_id","status","inference_result_id"],"title":"TaskStatusResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}